+++ components/mybot_platform/src/internal/board_actions.h
@@ -1,0 +1,15 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+#ifndef MYBOT_BOARD_ACTIONS_H_
+#define MYBOT_BOARD_ACTIONS_H_
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int mybot_board_handle_boot_long_press(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* MYBOT_BOARD_ACTIONS_H_ */